+++ SirensMoon/FX_AfterBurner.cpp
@@ -23,7 +23,7 @@ void FX_AfterBurner::Update() {
 }
 void FX_AfterBurner::Easing(int elapsed) {
 	auto linear = Easing::GetMode("Linear");
-	_animNo = linear(elapsed, 0, static_cast<int>(_cg.size()), _lifeTime);
+	_animNo = static_cast<int>(linear(elapsed, 0, static_cast<int>(_cg.size()), _lifeTime));
 
 }
 void FX_AfterBurner::CheckDeadOwner(){
