+++ SirensMoon/FX_BossFire.cpp
@@ -1,7 +1,6 @@
 #include "FX_BossFire.h"
 #include "Easing.h"
 
-
 FX_BossFire::FX_BossFire(Game& game, ModeGame& mode, const Vector2& pos, int startTime)
 	:Effect(game, mode, pos, startTime)
 {
@@ -15,5 +14,5 @@ FX_BossFire::FX_BossFire(Game& game, ModeGame& mode, const Vector2& pos, int sta
 
 void FX_BossFire::Easing(int elapsed) {
 	auto linear = Easing::GetMode("Linear");
-	_animNo = linear(elapsed, 0, static_cast<int>(_cg.size()), _lifeTime);
+	_animNo = static_cast<int>(linear(elapsed, 0, static_cast<int>(_cg.size()), _lifeTime));
 }
