+++ SirensMoon/FX_Chargenow.cpp
@@ -55,5 +55,5 @@ void FX_Chargenow::Update() {
 
 void FX_Chargenow::Easing(int elapsed) {
 	auto linear = Easing::GetMode("Linear");
-	_animNo = linear(elapsed, 0, static_cast<int>(_cg.size()), _lifeTime);
+	_animNo = static_cast<int>(linear(elapsed, 0, static_cast<int>(_cg.size()), _lifeTime));
 }
