+++ SirensMoon/WantedInfo.cpp
@@ -4,7 +4,7 @@
 class SplitWindow;
 
 WantedInfo::WantedInfo(Game& game, ModeBase& mode, SplitWindow& window, Vector2 pos, Vector2 size)
-	:UIBase{ game,mode,window,pos,size },_timer{0}
+	:UIBase{ game,mode,window,pos,size }, _timer{ 0 }
 {
 	_visible = false;
 	_cg.first = ImageServer::LoadGraph("resource/UI/WantedInfo/1.png");
@@ -16,14 +16,14 @@ void WantedInfo::Update() {
 		return;
 	}
 	++_timer;
-	
+
 }
 
 void WantedInfo::Render() {
 	if (_visible) {
 		if (_timer < 121) {
 			int cg{ -1 };
-			if (_timer % 60<30) {
+			if (_timer % 60 < 30) {
 				cg = _cg.second;
 			}
 			else {
@@ -34,21 +34,20 @@ void WantedInfo::Render() {
 				sizey = _timer * 50;
 			}
 			else {
-				sizey = (121-_timer) * 50;
+				sizey = (121 - _timer) * 50;
 			}
-			sizey=Math::Clamp(sizey,0,210);
-			DrawExtendGraph(_pos.x, _pos.y - sizey /2,
-				_pos.x+splitscreen_W, _pos.y + sizey /2,cg, 1);
+			sizey = Math::Clamp(sizey, 0, 210);
+			DrawExtendGraph(static_cast<int>(_pos.x), static_cast<int>(_pos.y - sizey / 2),
+				static_cast<int>(_pos.x + splitscreen_W), static_cast<int>(_pos.y + sizey / 2), cg, 1);
 		}
 		else {
-			DrawExtendGraph(_pos.x+splitscreen_W-290, 85,
-				_pos.x + splitscreen_W - 290+260,85+60,_cg.second, 1);
+			DrawExtendGraph(static_cast<int>(_pos.x + splitscreen_W - 290), 85,
+				static_cast<int>(_pos.x + splitscreen_W - 290 + 260), 85 + 60, _cg.second, 1);
 		}
-
 	}
 }
 
-void WantedInfo::TargetSpawnEvent(){
+void WantedInfo::TargetSpawnEvent() {
 	_visible = true;
 	_timer = 0;
 }
